Finished disassembly of server.js into common.js, cookies.js and session.js.
