+++ server/src/server.js
@@ -60,9 +60,14 @@ const pgQueryP_metered_readOnly = pg.pgQueryP_metered_readOnly;
 const pgQueryP_readOnly = pg.pgQueryP_readOnly;
 const pgQueryP_readOnly_wRetryIfEmpty = pg.pgQueryP_readOnly_wRetryIfEmpty;
 
-const User = require('./user');
+const cookies = require('./utils/cookies');
+const COOKIES = cookies.COOKIES;
+const COOKIES_TO_CLEAR = cookies.COOKIES_TO_CLEAR;
 
+const User = require('./user');
 const Conversation = require('./conversation');
+const Session = require('./session');
+const Utils = require('./utils/common');
 
 // # Slack setup
 
@@ -453,164 +458,20 @@ const sql_reports = sql.define({
 //     };
 // }());
 
-function encrypt(text) {
-  const algorithm = 'aes-256-ctr';
-  const password = process.env.ENCRYPTION_PASSWORD_00001;
-  const cipher = crypto.createCipher(algorithm, password);
-  var crypted = cipher.update(text,'utf8','hex');
-  crypted += cipher.final('hex');
-  return crypted;
-}
-
-function decrypt(text) {
-  const algorithm = 'aes-256-ctr';
-  const password = process.env.ENCRYPTION_PASSWORD_00001;
-  const decipher = crypto.createDecipher(algorithm, password);
-  var dec = decipher.update(text,'hex','utf8');
-  dec += decipher.final('utf8');
-  return dec;
-}
-decrypt; // appease linter
-function makeSessionToken() {
-  // These can probably be shortened at some point.
-  return crypto.randomBytes(32).toString('base64').replace(/[^A-Za-z0-9]/g, "").substr(0, 20);
-}
-
-// But we need to squeeze a bit more out of the db right now,
-// and generally remove sources of uncertainty about what makes
-// various queries slow. And having every single query talk to PG
-// adds a lot of variability across the board.
-const userTokenCache = new LruCache({
-  max: 9000,
-});
-
-function getUserInfoForSessionToken(sessionToken, res, cb) {
-  let cachedUid = userTokenCache.get(sessionToken);
-  if (cachedUid) {
-    cb(null, cachedUid);
-    return;
-  }
-  pgQuery("select uid from auth_tokens where token = ($1);", [sessionToken], function(err, results) {
-    if (err) {
-      console.error("token_fetch_error");
-      cb(500);
-      return;
-    }
-    if (!results || !results.rows || !results.rows.length) {
-      console.error("token_expired_or_missing");
-
-      cb(403);
-      return;
-    }
-    let uid = results.rows[0].uid;
-    userTokenCache.set(sessionToken, uid);
-    cb(null, uid);
-  });
-}
-function createPolisLtiToken(tool_consumer_instance_guid, lti_user_id) {
-  return ["xPolisLtiToken", tool_consumer_instance_guid, lti_user_id].join(":::");
-}
-
-function isPolisLtiToken(token) {
-  return token.match(/^xPolisLtiToken/);
-}
-function isPolisSlackTeamUserToken(token) {
-  return token.match(/^xPolisSlackTeamUserToken/);
-}
-
-// function sendSlackEvent(slack_team, o) {
-//   return pgQueryP("insert into slack_bot_events (slack_team, event) values ($1, $2);", [slack_team, o]);
-// }
-function sendSlackEvent(o) {
-  return pgQueryP("insert into slack_bot_events (event) values ($1);", [o]);
-}
-
-function parsePolisLtiToken(token) {
-  let parts = token.split(/:::/);
-  let o = {
-    // parts[0] === "xPolisLtiToken", don't need that
-    tool_consumer_instance_guid: parts[1],
-    lti_user_id: parts[2],
-  };
-  return o;
-}
-
-function getUserInfoForPolisLtiToken(token) {
-  let o = parsePolisLtiToken(token);
-  return pgQueryP("select uid from lti_users where tool_consumer_instance_guid = $1 and lti_user_id = $2", [
-    o.tool_consumer_instance_guid,
-    o.lti_user_id,
-  ]).then(function(rows) {
-    return rows[0].uid;
-  });
-}
-
-function startSession(uid, cb) {
-  let token = makeSessionToken();
-  //winston.log("info",'startSession: token will be: ' + sessionToken);
-  winston.log("info", 'startSession');
-  pgQuery("insert into auth_tokens (uid, token, created) values ($1, $2, default);", [uid, token], function(err, repliesSetToken) {
-    if (err) {
-      cb(err);
-      return;
-    }
-    winston.log("info", 'startSession: token set.');
-    cb(null, token);
-  });
-}
-
-function endSession(sessionToken, cb) {
-  pgQuery("delete from auth_tokens where token = ($1);", [sessionToken], function(err, results) {
-    if (err) {
-      cb(err);
-      return;
-    }
-    cb(null);
-  });
-}
-function setupPwReset(uid, cb) {
-  function makePwResetToken() {
-    // These can probably be shortened at some point.
-    return crypto.randomBytes(140).toString('base64').replace(/[^A-Za-z0-9]/g, "").substr(0, 100);
-  }
-  let token = makePwResetToken();
-  pgQuery("insert into pwreset_tokens (uid, token, created) values ($1, $2, default);", [uid, token], function(errSetToken, repliesSetToken) {
-    if (errSetToken) {
-      cb(errSetToken);
-      return;
-    }
-    cb(null, token);
-  });
-}
-
-function getUidForPwResetToken(pwresettoken, cb) {
-  // TODO "and created > timestamp - x"
-  pgQuery("select uid from pwreset_tokens where token = ($1);", [pwresettoken], function(errGetToken, results) {
-    if (errGetToken) {
-      console.error("pwresettoken_fetch_error");
-      cb(500);
-      return;
-    }
-    if (!results || !results.rows || !results.rows.length) {
-      console.error("token_expired_or_missing");
-      cb(403);
-      return;
-    }
-    cb(null, {
-      uid: results.rows[0].uid,
-    });
-  });
-}
-
-function clearPwResetToken(pwresettoken, cb) {
-  pgQuery("delete from pwreset_tokens where token = ($1);", [pwresettoken], function(errDelToken, repliesSetToken) {
-    if (errDelToken) {
-      cb(errDelToken);
-      return;
-    }
-    cb(null);
-  });
-}
+const encrypt = Session.encrypt;
+const decrypt = Session.decrypt;
+const makeSessionToken = Session.makeSessionToken;
+const getUserInfoForSessionToken = Session.getUserInfoForSessionToken;
+const createPolisLtiToken = Session.createPolisLtiToken;
+const isPolisLtiToken = Session.isPolisLtiToken;
+const isPolisSlackTeamUserToken = Session.isPolisSlackTeamUserToken;
+const sendSlackEvent = Session.sendSlackEvent;
+const getUserInfoForPolisLtiToken = Session.getUserInfoForPolisLtiToken;
+const startSession = Session.startSession;
+const endSession = Session.endSession;
+const setupPwReset = Session.setupPwReset;
+const getUidForPwResetToken = Session.getUidForPwResetToken;
+const clearPwResetToken = Session.clearPwResetToken;
 
 function hasAuthToken(req) {
   return !!req.cookies[COOKIES.TOKEN];
@@ -778,49 +639,7 @@ String.prototype.hashCode = function() {
   return hash;
 };
 
-const COOKIES = {
-  COOKIE_TEST: 'ct',
-  HAS_EMAIL: 'e',
-  TOKEN: 'token2',
-  UID: 'uid2',
-  REFERRER: 'ref',
-  PARENT_REFERRER: 'referrer',
-  PARENT_URL: 'parent_url',
-  USER_CREATED_TIMESTAMP: 'uc',
-  PERMANENT_COOKIE: 'pc',
-  TRY_COOKIE: 'tryCookie',
-  PLAN_NUMBER: 'plan', // not set if trial user
-};
-const COOKIES_TO_CLEAR = {
-  e: true,
-  token2: true,
-  uid2: true,
-  uc: true,
-  plan: true,
-  referrer: true,
-  parent_url: true,
-};
-
 function initializePolisHelpers() {
-  let polisTypes = {
-    reactions: {
-      push: 1,
-      pull: -1,
-      see: 0,
-      pass: 0,
-    },
-    staractions: {
-      unstar: 0,
-      star: 1,
-    },
-    mod: {
-      ban: -1,
-      unmoderated: 0,
-      ok: 1,
-    },
-  };
-  polisTypes.reactionValues = _.values(polisTypes.reactions);
-  polisTypes.starValues = _.values(polisTypes.staractions);
   // // If there are any comments which have no votes by the owner, create a PASS vote by the owner.
   // pgQuery("select * from comments", [], function(err, comments) {
   //     pgQuery("select * from votes", [], function(err, votes) {
@@ -861,123 +680,6 @@ function initializePolisHelpers() {
   //     });
   // });
 
-  let oneYear = 1000 * 60 * 60 * 24 * 365;
-
-  function setCookie(req, res, setOnPolisDomain, name, value, options) {
-    let o = _.clone(options || {});
-    o.path = _.isUndefined(o.path) ? '/' : o.path;
-    o.maxAge = _.isUndefined(o.maxAge) ? oneYear : o.maxAge;
-    if (setOnPolisDomain) {
-      o.secure = _.isUndefined(o.secure) ? true : o.secure;
-      o.domain = _.isUndefined(o.domain) ? '.pol.is' : o.domain;
-      // if (/pol.is/.test(req.headers.host)) {
-      //   o.domain = '.pol.is';
-      // }
-    }
-    res.cookie(name, value, o);
-  }
-
-  function setParentReferrerCookie(req, res, setOnPolisDomain, referrer) {
-    setCookie(req, res, setOnPolisDomain, COOKIES.PARENT_REFERRER, referrer, {
-      httpOnly: true,
-    });
-  }
-
-  function setParentUrlCookie(req, res, setOnPolisDomain, parent_url) {
-    setCookie(req, res, setOnPolisDomain, COOKIES.PARENT_URL, parent_url, {
-      httpOnly: true,
-    });
-  }
-
-  function setPlanCookie(req, res, setOnPolisDomain, planNumber) {
-    if (planNumber > 0) {
-      setCookie(req, res, setOnPolisDomain, COOKIES.PLAN_NUMBER, planNumber, {
-        // not httpOnly - needed by JS
-      });
-    }
-    // else falsy
-
-  }
-
-  function setHasEmailCookie(req, res, setOnPolisDomain, email) {
-    if (email) {
-      setCookie(req, res, setOnPolisDomain, COOKIES.HAS_EMAIL, 1, {
-        // not httpOnly - needed by JS
-      });
-    }
-    // else falsy
-  }
-
-  function setUserCreatedTimestampCookie(req, res, setOnPolisDomain, timestamp) {
-    setCookie(req, res, setOnPolisDomain, COOKIES.USER_CREATED_TIMESTAMP, timestamp, {
-      // not httpOnly - needed by JS
-    });
-  }
-
-  function setTokenCookie(req, res, setOnPolisDomain, token) {
-    setCookie(req, res, setOnPolisDomain, COOKIES.TOKEN, token, {
-      httpOnly: true,
-    });
-  }
-
-  function setUidCookie(req, res, setOnPolisDomain, uid) {
-    setCookie(req, res, setOnPolisDomain, COOKIES.UID, uid, {
-      // not httpOnly - needed by JS
-    });
-  }
-
-  function setPermanentCookie(req, res, setOnPolisDomain, token) {
-    setCookie(req, res, setOnPolisDomain, COOKIES.PERMANENT_COOKIE, token, {
-      httpOnly: true,
-    });
-  }
-
-  function setCookieTestCookie(req, res, setOnPolisDomain) {
-    setCookie(req, res, setOnPolisDomain, COOKIES.COOKIE_TEST, 1, {
-      // not httpOnly - needed by JS
-    });
-  }
-
-  function shouldSetCookieOnPolisDomain(req) {
-    let setOnPolisDomain = !domainOverride;
-    let origin = req.headers.origin || "";
-    if (setOnPolisDomain && origin.match(/^http:\/\/localhost:[0-9]{4}/)) {
-      setOnPolisDomain = false;
-    }
-    return setOnPolisDomain;
-  }
-
-  function addCookies(req, res, token, uid) {
-    return getUserInfoForUid2(uid).then(function(o) {
-      let email = o.email;
-      let created = o.created;
-      let plan = o.plan;
-
-      let setOnPolisDomain = shouldSetCookieOnPolisDomain(req);
-
-      setTokenCookie(req, res, setOnPolisDomain, token);
-      setUidCookie(req, res, setOnPolisDomain, uid);
-      setPlanCookie(req, res, setOnPolisDomain, plan);
-      setHasEmailCookie(req, res, setOnPolisDomain, email);
-      setUserCreatedTimestampCookie(req, res, setOnPolisDomain, created);
-      if (!req.cookies[COOKIES.PERMANENT_COOKIE]) {
-        setPermanentCookie(req, res, setOnPolisDomain, makeSessionToken());
-      }
-      res.header("x-polis", token);
-    });
-  }
-
-  function getPermanentCookieAndEnsureItIsSet(req, res) {
-    let setOnPolisDomain = shouldSetCookieOnPolisDomain(req);
-    if (!req.cookies[COOKIES.PERMANENT_COOKIE]) {
-      let token = makeSessionToken();
-      setPermanentCookie(req, res, setOnPolisDomain, token);
-      return token;
-    } else {
-      return req.cookies[COOKIES.PERMANENT_COOKIE];
-    }
-  }
-
   function generateHashedPassword(password, callback) {
     bcrypt.genSalt(12, function(errSalt, salt) {
       if (errSalt) {
@@ -991,6 +693,17 @@ function initializePolisHelpers() {
       });
     });
   }
+
+  const polisTypes = Utils.polisTypes;
+  const setCookie = cookies.setCookie;
+  const setParentReferrerCookie = cookies.setParentReferrerCookie;
+  const setParentUrlCookie = cookies.setParentUrlCookie;
+  const setPlanCookie = cookies.setPlanCookie;
+  const setPermanentCookie = cookies.setPermanentCookie;
+  const setCookieTestCookie = cookies.setCookieTestCookie;
+  const shouldSetCookieOnPolisDomain = cookies.shouldSetCookieOnPolisDomain;
+  const addCookies = cookies.addCookies;
+  const getPermanentCookieAndEnsureItIsSet = cookies.getPermanentCookieAndEnsureItIsSet;
 
   const pidCache = User.pidCache;
   const getPid = User.getPid;
@@ -1593,26 +1306,8 @@ function initializePolisHelpers() {
   ////////////////////////////////////////////
   ////////////////////////////////////////////
 
-  function strToHex(str) {
-    let hex, i;
-    // let str = "\u6f22\u5b57"; // "\u6f22\u5b57" === "漢字"
-    let result = "";
-    for (i = 0; i < str.length; i++) {
-      hex = str.charCodeAt(i).toString(16);
-      result += ("000" + hex).slice(-4);
-    }
-    return result;
-  }
-
-  function hexToStr(hexString) {
-    let j;
-    let hexes = hexString.match(/.{1,4}/g) || [];
-    let str = "";
-    for (j = 0; j < hexes.length; j++) {
-      str += String.fromCharCode(parseInt(hexes[j], 16));
-    }
-    return str;
-  }
+  const strToHex = Utils.strToHex;
+  const hexToStr = Utils.hexToStr;
 
   function handle_GET_launchPrep(req, res) {
     let setOnPolisDomain = !domainOverride;
@@ -4005,7 +3700,7 @@ Email verified! You can close this tab or hit the back button.
     });
   }
 
-  const addLtiUserifNeeded = User.addLtiUserIfNeeded;
+  const addLtiUserIfNeeded = User.addLtiUserIfNeeded;
   const addLtiContextMembership = User.addLtiContextMembership;
 
   function subscribeToNotifications(zid, uid, email) {
@@ -12795,6 +12490,7 @@ CREATE TABLE slack_user_invites (
     }, 100);
 
     doGetConversationPreloadInfo(conversation_id).then(function(x) {
+      console.log(x);
       let preloadData = {
         conversation: x,
         // Nothing user-specific can go here, since we want to cache these per-conv index files on the CDN.
